+++ cypress/integration/checkout/checkout.multipiece.items.spec.js
@@ -8,7 +8,7 @@ import Helper from '../../support/fragments/finance/financeHelper';
 import CheckOutActions from '../../support/fragments/check-out-actions/check-out-actions';
 import NewServicePoint from '../../support/fragments/service_point/newServicePoint';
 import ServicePoints from '../../support/fragments/settings/tenant/servicePoints';
-import ConfirmMultiplePiecesItemCheckOut from '../../support/fragments/checkout/confirmMultiplePiecesItemCheckOut';
+import ConfirmMultipieceCheckOutModal from '../../support/fragments/checkout/confirmMultipieceCheckOutModal';
 import SwitchServicePoint from '../../support/fragments/service_point/switchServicePoint';
 
 describe('Check Out', () => {
@@ -130,19 +130,20 @@ describe('Check Out', () => {
     CheckOutActions.checkIsInterfacesOpened();
     CheckOutActions.checkOutItem(userBarcode, FIRST_ITEM_BARCODE);
     CheckOutActions.checkPatronInformation(user.username, userBarcode);
+    CheckOutActions.checkConfirmMultipieceCheckOutModal();
     CheckOutActions.checkOutItem(userBarcode, SECOND_ITEM_BARCODE);
-    ConfirmMultiplePiecesItemCheckOut.checkIsModalConsistOf(instanceTitle, quantityPiecesForSecondItem, descriptionOfPiece);
-    ConfirmMultiplePiecesItemCheckOut.cancelModal();
+    ConfirmMultipieceCheckOutModal.checkIsModalConsistOf(instanceTitle, quantityPiecesForSecondItem, descriptionOfPiece);
+    ConfirmMultipieceCheckOutModal.cancelModal();
     CheckOutActions.checkOutItem(userBarcode, SECOND_ITEM_BARCODE);
-    ConfirmMultiplePiecesItemCheckOut.checkIsModalConsistOf(instanceTitle, quantityPiecesForSecondItem, descriptionOfPiece);
-    ConfirmMultiplePiecesItemCheckOut.confirmMultiplePiecesItemModal();
+    ConfirmMultipieceCheckOutModal.checkIsModalConsistOf(instanceTitle, quantityPiecesForSecondItem, descriptionOfPiece);
+    ConfirmMultipieceCheckOutModal.confirmMultiplePiecesItemModal();
     CheckOutActions.checkOutItem(userBarcode, THIRD_ITEM_BARCODE);
-    ConfirmMultiplePiecesItemCheckOut.checkIsModalConsistOf(instanceTitle, quantityPiecesForThirdItem, descriptionOfPiece);
-    ConfirmMultiplePiecesItemCheckOut.checkMissingPiecesInModal(quantityOfMissingPieces, missingPieceDescription);
-    ConfirmMultiplePiecesItemCheckOut.confirmMultiplePiecesItemModal();
+    ConfirmMultipieceCheckOutModal.checkIsModalConsistOf(instanceTitle, quantityPiecesForThirdItem, descriptionOfPiece);
+    ConfirmMultipieceCheckOutModal.checkMissingPiecesInModal(quantityOfMissingPieces, missingPieceDescription);
+    ConfirmMultipieceCheckOutModal.confirmMultiplePiecesItemModal();
     CheckOutActions.checkOutItem(userBarcode, FOURTH_ITEM_BARCODE);
-    ConfirmMultiplePiecesItemCheckOut.checkIsModalConsistOf(instanceTitle, quantityPiecesForFourthItem, dash);
-    ConfirmMultiplePiecesItemCheckOut.checkMissingPiecesInModal(quantityOfMissingPieces, missingPieceDescription);
-    ConfirmMultiplePiecesItemCheckOut.confirmMultiplePiecesItemModal();
+    ConfirmMultipieceCheckOutModal.checkIsModalConsistOf(instanceTitle, quantityPiecesForFourthItem, dash);
+    ConfirmMultipieceCheckOutModal.checkMissingPiecesInModal(quantityOfMissingPieces, missingPieceDescription);
+    ConfirmMultipieceCheckOutModal.confirmMultiplePiecesItemModal();
   });
 });
