+++ cypress/support/fragments/checkout/confirmMultipieceCheckOutModal.js
